chore: CalculatorItemQueue타입을 Model Group으로 이동 #2
